backend/pyproject.toml
[tool.poetry]
name = "modernreader-backend"
version = "0.1.0"
description = "CARE-aligned multisensory reading companion API"
authors = ["ModernReader Team <team@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.28"
psycopg = {extras = ["binary"], version = "^3.1.18"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
pydantic-settings = "^2.2.0"
email-validator = "^2.3.0"
openai = "^2.6.1"
anthropic = "^0.72.0"
langchain = "^1.0.3"
langchain-openai = "^1.0.1"
langchain-anthropic = "^1.0.1"
langchain-google-genai = "^3.0.0"
chromadb = "^1.3.0"
sentence-transformers = "^5.1.2"
langchain-community = "^0.4.1"
langchain-text-splitters = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
httpx = "^0.27.0"
ruff = "^0.3.0"
mypy = "^1.18.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
